게시글 수정, 파일 유효성 검사 구현

+++ src/main/java/com/sparta/newspeed/user/service/UserService.java
@@ -47,8 +47,8 @@ public class UserService {
     public UserInfoUpdateDto updateUser(Long userSeq, UserInfoUpdateDto requestDto, MultipartFile file) {
         User user = findById(userSeq);
         if (file != null) {
-            s3Service.deleteFile(user.getPhotoName());
-            user.setPhotoName(s3Service.uploadFile(file));
+            if(user.getPhotoName() != null) { s3Service.deleteFile(user.getPhotoName()); }
+            user.setPhotoName(s3Service.uploadFile(file, "profile"));
         }
         user.updateUserInfo(requestDto);
         return new UserInfoUpdateDto(user);
